+++ tests/ekobox/test_gen_info.py
@@ -2,14 +2,12 @@
 import math
 
 import numpy as np
-import pytest
 
 from ekobox import gen_info as g_i
 from ekobox import gen_op as g_o
 from ekobox import gen_theory as g_t
 
 
-@pytest.mark.isolated
 def benchmark_create_info_file():
     op = g_o.gen_op_card([10, 100])
     theory = g_t.gen_theory_card(1, 10.0, update={"alphas": 0.2})
