+++ tests/ekobox/test_gen_theory.py
@@ -4,7 +4,6 @@ import pytest
 from ekobox import gen_theory as g_t
 
 
-@pytest.mark.isolated
 def benchmark_gen_theory_card():
     theory = g_t.gen_theory_card(0, 1.0)
     assert theory["PTO"] == 0
@@ -19,7 +18,6 @@ def benchmark_gen_theory_card():
     assert theory["mb"] == 132.3
 
 
-@pytest.mark.isolated
 def benchmark_export_load_theory_card(tmp_path, cd):
     with cd(tmp_path):
         theory = g_t.gen_theory_card(2, 12.3, name="debug_theory")
